requirements.txt
rich
pandas
python-dateutil
